requirements.txt
langchain_cohere >= 0.1.7
langchain_openai >= 0.1.6

pyperclip >= 1.3.0

langchain == 0.2.2

beautifulsoup4 >= 4.10.0
faiss-cpu >= 1.8.0

requests >= 2.31.0

sentence-transformers >= 3.0.1
psutil~=6.0.0